refactor: move event detail form to component
